packages/cli/pyproject.toml
[build-system]
requires = ["packaging>=22", "tomli"]
build-backend = "pymonorepo.backend"
backend-path = ["../build/src"]

[project]
name = "pymonorepo-cli"
dynamic = [
    "version",
    "description",
]
authors = [{name = "Chris Sewell", email = "chrisj_sewell@hotmail.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "click",
    "pyyaml",
    "pymonorepo",
]
requires-python = ">=3.7"

[project.scripts]
pmr = "pymonorepo_cli:__main__"
